made auth blueprint
