+++ app.py
@@ -13,10 +13,10 @@ from models.user import User, UserSchema
 from models.book import Book, BookSchema
 from models.comment import Comment, CommentSchema
 from models.review import Review, ReviewSchema
-import psycopg2
 from init import db, ma, bcrypt, jwt
 
 from blueprints.cli_bp import cli_bp
+from blueprints.auth_bp import auth_bp
 
 app = Flask(__name__)
 
@@ -29,14 +29,7 @@ jwt.init_app(app)
 bcrypt.init_app(app)
 
 app.register_blueprint(cli_bp)
-    
-def admin_or_owner_required(owner_email):
-    user_email = get_jwt_identity()
-    stmt = db.select(User).filter_by(email=user_email)
-    user = db.session.scalar(stmt)
-    if not (user and (user.is_admin or user_email == owner_email)):
-        abort(401, description='You must be an admin or the owner')
-
+app.register_blueprint(auth_bp)
     
 @app.route("/books", methods=["GET"])
 def books():
@@ -44,33 +37,6 @@ def books():
     result = books_schema.dump(books_list)
     return jsonify(result)
 
-@app.route("/register", methods=['POST'])
-def register():
-    email = request.form['email']
-    test = User.query.filter_by(email=email).first()
-    if test: 
-        return jsonify(message="that email already exists"), 409
-    else:
-        username = request.form['username']
-        email = request.form['email']
-        password = request.form['password']
-        user = User(username=username, email=email, password=bcrypt.generate_password_hash(password).decode('utf8'))
-        db.session.add(user)
-        db.session.commit()
-        return UserSchema(exclude=['password']).dump(user), 201
-    
-@app.route('/login', methods=['POST'])
-def login():
-    try:
-        stmt = db.select(User).filter_by(email=request.json['email'])
-        user = db.session.scalar(stmt)
-        if user and bcrypt.check_password_hash(user.password, request.json['password']):
-            token = create_access_token(identity=user.email, expires_delta=timedelta(days=1))
-            return {'token': token, 'user': UserSchema(exclude=['password']).dump(user)}
-        else:
-            return {'error': 'Invalid email address or password'}, 401
-    except KeyError:
-        return {'error': 'Email and password are required'}, 400
 
 
 @app.route('/book_details/<int:book_id>', methods=['GET'])
